refactor(api): split main.py into composition + middleware + handlers (CHAOS-1543)
src/dev_health_ops/api/main.py was 1563 lines mixing FastAPI composition,
middleware wiring, exception handlers, lifespan, observability, and route
handlers. Extracted the cross-cutting concerns into focused private modules
so main.py is closer to pure composition. Pure refactor: zero behavior
change, no test modifications, app import path stable.

New modules:
- api/_errors.py:        exception handlers + register_exception_handlers(app)
- api/_middleware.py:    middleware wiring + register_middleware(app)
                         (CORS, SlowAPI, security headers, OriginValidation,
                          OrgId, Impersonation, CorrelationId, GraphQL size)
- api/_lifespan.py:      lifespan context manager (licensing + bundle
                         validation startup, ClickHouse client shutdown)
- api/_observability.py: register_observability(app) (OpenTelemetry +
                         optional Prometheus /metrics)
- api/_health.py:        DB-URL resolution + health-check helpers
                         (_db_url, _postgres_url, _clickhouse_url,
                          _analytics_db_url, _check_*_health)

main.py changes:
- Helper functions, exception handlers, middleware wiring, lifespan,
  Prometheus block, and OpenTelemetry instrumentation removed and
  replaced by register_* calls.
- Names that tests monkeypatch (e.g. _check_redis_health, _analytics_db_url,
  _generic_exception_handler) re-imported in main.py so existing patch paths
  continue to work unchanged.
- Removed pre-existing `# type: ignore[arg-type]` on the validation handler
  by typing the parameter as Exception and adding a defensive isinstance
  check (behavior identical, types now correct).
- main.py drops from 1569 -> 1309 lines. Route handlers remain in main.py
  to preserve the long list of test monkeypatch targets
  (build_home_response, build_explain_response, etc.); extracting them into
  routers is a follow-up that requires updating those patch paths.

Verification:
- All 530 tests in tests/api/ + ingest/admin/redis/people/quadrant suites
  pass with redis running locally.
- python -c "from dev_health_ops.api.main import app; print(len(app.routes))"
  -> 204 routes.
- LSP diagnostics clean on every changed file.
- Ruff F/E checks clean on all new modules and main.py (the one remaining
  E501 is a pre-existing comment in an untouched route handler).
